corrige les tests unitaires
