+++ src/test/java/com/payline/payment/google/pay/service/impl/PaymentServiceImplTest.java
@@ -44,7 +44,7 @@ public class PaymentServiceImplTest {
 
     public void paymentRequest() throws GeneralSecurityException {
         PaymentRequest request = Utils.createCompletePaymentBuilder().build();
-        request.getPartnerConfiguration().getSensitiveProperties().put(PAYMENT_REQUEST_PAYMENT_DATA_KEY, GOOD_PAYMENT_DATA);
+        request.getPaymentFormContext().getPaymentFormParameter().put(PAYMENT_REQUEST_PAYMENT_DATA_KEY, GOOD_PAYMENT_DATA);
 
         doReturn(GOOD_RESPONSE_DATA).when(service).getDecryptedData(anyString(), anyString(), anyBoolean());
 
@@ -83,7 +83,7 @@ public class PaymentServiceImplTest {
     @Test
     public void paymentRequestKO() throws GeneralSecurityException {
         PaymentRequest request = Utils.createCompletePaymentBuilder().build();
-        request.getPartnerConfiguration().getSensitiveProperties().put(PAYMENT_REQUEST_PAYMENT_DATA_KEY, GOOD_PAYMENT_DATA);
+        request.getPaymentFormContext().getPaymentFormParameter().put(PAYMENT_REQUEST_PAYMENT_DATA_KEY, GOOD_PAYMENT_DATA);
 
 
         doThrow(new GeneralSecurityException()).when(service).getDecryptedData(anyString(), anyString(), anyBoolean());
