+++ src/test/resources/index.html
@@ -6,7 +6,7 @@
 	<meta content="utf-8" http-equiv="encoding">
 </head>
 
-<body id="body">
+<body id="paylineGooglePayBtnDivContainer">
 	<script type="text/javascript" src="http://code.jquery.com/jquery-1.7.1.min.js"></script>
 	<script type="text/javascript" src="init.js"></script>
 	<script type="text/javascript" src="https://pay.google.com/gp/p/js/pay.js"></script>
